#1 Move breakpoints, fix some styling in contact us
